Add sub-interpreter testing ground with unit test support
Since Python's API is in C, this "testing ground" (or whatever you'd
like to call it) is in C++ to make quick prototyping easier.

Unit tests are also made possible thanks to doctest.h - currently,
there's only an example test in there.

Note that this testing environment is by no means final and will
continue to evolve as more needs arise.

Signed-off-by: Max Carrara <m.carrara@proxmox.com>